requirements.txt
opencv-python==4.8.0.76
face-recognition==1.3.0
pandas==2.0.3
numpy==1.24.3
Pillow==10.0.0
flask==2.3.3
flask-cors==4.0.0
dlib==19.24.2
cmake==3.27.4